Upgrade
* Replace `blake2b.js` with the npm version
* Convert the project to `typescript`
* Update the `readme.md` file
