+++ .gitignore
@@ -1,2 +1,2 @@
-dist
+.idea
 node_modules
