+++ readme.md
@@ -1,4 +1,19 @@
-Build (will create a bundle and copy it to /tmp/handshake-util.js):
+# handshake-util
 
-    npm install
-    npm run build
+A collection of utilities for working with Handshake.
+
+
+## Installation
+
+```bash
+$ npm install handshake-util
+```
+
+## Build Browserify
+
+It will create a bundle and copy it to `/tmp/handshake-util.js`
+
+```bash
+$ npm install
+$ npm run build:browserify
+```
